Resubmit Stage 1 following comments
Addresses following comments:

1) You need to use <h#> headers in these titles. Use an <h1> for the very first main title, and <h2> to <h6> for the rest of titles or subtitles.
2) Oops! This image link is not working, since it pointing you tour local filesystem. You need to upload the image within your project, and use a relative path, like: images/me.jpg, for example.
3) I think you did not get the right idea in this part.
The purpose of classes here is not to describe the content. The purpose of the class is to define a style for a type of content.

For example, I could define three classes: 'box', title' and 'description'. Then I will create different contents, all of them with class 'title' for the content title and 'description' for the longer part of text, with a box surrounding each of them:

<div class="box">
    <h2 class="title">
        HTML Tags
    </h2>
    <p class="description">
        The content of this section goes here
    </p>
</div>

<div class="box">
    <h2 class="title">
        How to view Web page structure
    </h2>
    <p class="description">
        The content of this different section goes here
    </p>
</div>
You should use the same class (in the example above, 'title', for example) in more than one element.

Give it a try!